Get ostree repo config saving working-ish
Add a OstreeConfigRepoConfigFileSave that persists
a OstreeConfig to a repo_file.RepoFile

Add a clear_remote() and a set_remote to
repo_file.RepoFile

Change internal data of OstreeRemotes to a set
instead of a list.

Let OstreeConfigControler update it's remote_set
directly instead of via a OstreeRemotesUpdater
Need to revisit since remote section updating
is unknown.

OstreeConfigController also calls save() directly
on it's OstreeConfigs' RepoFile