Add a PluginManager.runiter() method.
This allows iteratting over an iterable of PluginManagerRun
objects. A PluginManagerRun will be created for every plugin
method that is a hook for the given slot. Ie, if 10 plugins
all register a 'content_update_hook', this lets you iterate
over all if you need. PluginManagerRun has attributes for
the associated conduit and plugin method.

ContentActionClient will uses this to wrap each content
plugin invocation to track their update reports.