Make OstreeConfig use OstreeConfigFileStore
If OstreeConfig hasn't created a repo_file_store
backing store, create one (a OstreeConfigFileStore)
and save()

Make OstreeConfigFileStore open config file
and create a RepoFile() if it doesn't have one,
and save().

This is to support creating an OstreeConfig
with a backing store, but being able to save
to the default.

Misc name cleanup.