Remove inj from 'api' module.
Content plugins that need entitlement info to
find Content info now should use an EntitlementSource
available on the ContentUpdateConduit as the
'ent_source' attribute.

EntitlementSource is just an iterable of Entitlements
(currently, that means the same thing as a
list_valid() on the injected ENT_DIR).

Plugins can currently assume that each item in
conduit.ent_source iterable with have a
'content' attribute that will be a list of
objects that looks like rhsm.certificate.Content

The current only implementation of EntitlementSource
is EntitltementDirEntitlementSource, based on the
inj'ed ENT_DIR.

Note the list of entitlements are calculated before
running any of the content update plugins, and does
not reflect if it changes as side effect of the
plugin hooks.

plugin/ostree/* no longer references api module
directly.