Add models module, for objects we pass to plugins
Move EntitlementSource to models.

Make EntitlementDirEntitlementSource a container
of models.Entitlement's (or models.EntitlementCertEntitlement
for our case).

EntitlementSource is an abstract version of EntitlementDirectory,
and Entitlement is a abstract version of a EntitlementCertificate.
EntitledContents is a abstract version of
EntitlementCertificate.content (a list of rhsm.certificate2.Content
 currently).

Add test cases for models.